requirements_dev.txt
asgiref==3.6.0
Django==3.2.19
django-cors-headers==3.11.0
django-decouple==2.1
djangorestframework==3.13.1
Pillow==9.5.0
psycopg2-binary==2.9.3
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
